+++ pistomp/handler.py
@@ -16,6 +16,12 @@
 
 from __future__ import annotations
 
+from typing import Any
+
+import common.token as Token
+from pistomp.analogmidicontrol import AnalogMidiControl
+from pistomp.encodermidicontrol import EncoderMidiControl
+from pistomp.footswitch import Footswitch
 from pistomp.tuner.source import TunerSourceFactory
 
 
@@ -23,7 +29,8 @@ class Handler:
     def __init__(self):
         self.homedir = None
         self.lcd = None
-        pass
+        self.hardware: Any = None
+        self.current: Any = None
 
     @property
     def lcd_poll_divisor(self) -> int:
@@ -104,3 +111,53 @@ class Handler:
 
     def set_tuner_source_factory(self, factory: "TunerSourceFactory") -> None:
         pass
+
+    #
+    # MIDI binding (shared by v1/v3 handlers)
+    #
+    def _apply_midi_binding(self, instance, symbol, binding):
+        # A MIDI mapping was learned in mod-ui. Update the matching parameter's
+        # binding and wire its hardware controller so the LCD reflects it without
+        # a pedalboard reload. Idempotent: replayed connect-dump maps are no-ops.
+        if self.current is None:
+            return
+        plugin = next((p for p in self.current.pedalboard.plugins
+                       if p is not None and p.instance_id == instance), None)
+        if plugin is None or plugin.parameters is None:
+            return
+        param = plugin.parameters.get(symbol)
+        if param is None or param.binding == binding:
+            return
+        controller = self.hardware.controllers.get(binding)
+        if controller is None:
+            return
+        param.binding = binding
+        is_footswitch = self._bind_controller_to_param(plugin, param, controller)
+        self._redraw_after_binding(controller, is_footswitch)
+
+    def _bind_controller_to_param(self, plugin, param, controller) -> bool:
+        # Wire a hardware controller to a plugin parameter. Returns True if the
+        # controller is a footswitch (so callers can track footswitch plugins).
+        # TODO possibly use a setter instead of accessing var directly
+        # What if multiple params could map to the same controller?
+        controller.parameter = param  # pyright: ignore[reportAttributeAccessIssue]
+        controller.set_value(param.value)
+        if controller not in plugin.controllers:
+            plugin.controllers.append(controller)
+        if isinstance(controller, Footswitch):
+            # TODO sort this list so selection orders correctly (sort on midi_CC?)
+            plugin.has_footswitch = True
+            controller.set_category(plugin.category)
+            return True
+        elif isinstance(controller, (AnalogMidiControl, EncoderMidiControl)):
+            key = "%s:%s" % (plugin.instance_id, param.name)
+            controller.cfg[Token.CATEGORY] = plugin.category  # somewhat LAME adding to cfg dict
+            controller.cfg[Token.TYPE] = controller.type
+            controller.cfg[Token.ID] = controller.id
+            self.current.analog_controllers[key] = controller.cfg
+        return False
+
+    def _redraw_after_binding(self, controller, is_footswitch):
+        # Refresh the LCD after a learned binding. Subclasses redraw at their
+        # own granularity.
+        raise NotImplementedError()
